fix: rename components AuthButton
